feat: add functionality for resyncing accounts
