+++ packages/shared/lib/wallet.ts
@@ -6,6 +6,7 @@ import type { Event, TransactionEventPayload, ConfirmationStateChangeEventPayloa
 import { mnemonic } from 'shared/lib/app'
 import { activeProfile, updateProfile } from 'shared/lib/profile'
 import { _ } from 'shared/lib/i18n'
+import type { SyncedAccount } from './typings/account'
 
 export const WALLET_STORAGE_DIRECTORY = '__storage__'
 
@@ -245,3 +246,47 @@ export const getLatestMessages = (accounts: Account[], count = 10): Message[] =>
         })
         .slice(0, count)
 }
+
+/**
+ * Updates accounts information after a successful sync accounts operation
+ * 
+ * @method updateAccounts
+ * 
+ * @param {SyncedAccount[]} syncedAccounts
+ * 
+ * @returns {void} 
+ */
+export const updateAccounts = (syncedAccounts: SyncedAccount[]): void => {
+    const _update = (existingPayload, newPayload, prop) => {
+        const existingPayloadMap = existingPayload.reduce((acc, object) => {
+            acc[object[prop]] = object
+
+            return acc
+        }, {})
+
+        const newPayloadMap = newPayload.reduce((acc, object) => {
+            acc[object[prop]] = object
+
+            return acc
+        }, {})
+
+        return Object.values(Object.assign({}, existingPayloadMap, newPayloadMap))
+    }
+
+    const { accounts } = get(wallet)
+
+    accounts.update((storedAccounts) => {
+        return storedAccounts.map((storedAccount) => {
+            const syncedAccount = syncedAccounts.find((_account) => _account.id === storedAccount.id)
+
+            return Object.assign({}, storedAccount, {
+                // Update deposit address
+                depositAddress: syncedAccount.depositAddress.address,
+                // If we have received a new address, simply add it;
+                // If we have received an existing address, update the properties.
+                addresses: _update(storedAccount.addresses, syncedAccount.addresses, 'address'),
+                messages: _update(storedAccount.messages, syncedAccount.messages, 'id'),
+            })
+        })
+    })
+};
